Chagne file structure
